requirements.txt
pandas==2.2.2
nbconvert==7.16.4
ipykernel==6.29.4
jupyter==1.0.0
ipykernel==6.29.4
ipywidgets==8.1.3
matplotlib==3.9.0
numpy==1.24.1
rioxarray==0.15.0
dvc==3.51.2
ensure==1.0.4
python-box==7.1.1
zenodo-get==1.6.1
scikit-learn==1.5.1